Rename grpc files
