+++ package.json
@@ -9,7 +9,7 @@
     "lint": "eslint src/ tests/",
     "lint:fix": "eslint --fix .",
     "test": "jest",
-    "compile": "grpc_tools_node_protoc --proto_path=src/typeDefs --js_out=import_style=commonjs,binary:src/lib src/typeDefs/resolveError.proto"
+    "compile": "grpc_tools_node_protoc --proto_path=src/typeDefs --js_out=import_style=commonjs,binary:src/lib src/typeDefs/resolve_error.proto"
   },
   "author": "MrOrz",
   "license": "MIT",
