+++ src/typeDefs/url_resolver.proto
@@ -1,8 +1,8 @@
 syntax = "proto3";
 
-import "resolveError.proto";
+import "resolve_error.proto";
 
-package urlResolver;
+package url_resolver;
 
 service UrlResolver {
   rpc ResolveUrl (UrlsRequest) returns (stream UrlReply) {}
@@ -21,8 +21,11 @@ message UrlReply {
   string canonical = 2;
   string title = 3;
   string summary = 4;
-  string topImageUrl = 5;
+  string top_image_url = 5;
   string html = 6;
   string status = 7;
-  ResolveError error = 8;
+  oneof has_error {
+    ResolveError error = 8;
+    bool successfully_resolved = 9;
+  }
 }
